actions/release-notification/requirements.txt
requests
pygithub
packaging